+++ src/source_map_support.js
@@ -1,0 +1,3 @@
+/* globals SourceMapSupport: false */
+
+SourceMapSupport.install();